split error defs into separate file
